pyproject.toml
[project]
name = "django-headless"
version = "0.1.0"
description = "Turn your Django app into a headless CMS."
authors = [
    {name = "Leon van der Grient",email = "leon@devtastic.io"}
]
license = {text = "BSD 3-Clause License"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "rich (>=13.9.4,<14.0.0)",
    "drf-flex-fields (>=1.0.2,<2.0.0)",
    "jsonschema (>=4.23.0,<5.0.0)",
    "pyhumps (>=3.8.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "djangorestframework-camel-case (>=1.4.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.development.dependencies]
black = "^25.1.0"
djangorestframework = "^3.15.2"
django = "^5.1.5"

